Login, Authentication, Crud

+++ src/Controller/EntrepriseController.php
@@ -3,45 +3,76 @@
 namespace App\Controller;
 
 use App\Entity\Entreprise;
+use App\Form\EntrepriseType;
 use App\Repository\EntrepriseRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
-#[Route('/entreprise', name: 'app_entreprise_')]
+#[Route('/entreprise')]
 class EntrepriseController extends AbstractController
 {
-    #[Route('/', name: 'index')]
+    #[Route('/', name: 'app_entreprise_index', methods: ['GET'])]
     public function index(EntrepriseRepository $entrepriseRepository): Response
     {
-        $entreprises = $entrepriseRepository->findAll();
         return $this->render('entreprise/index.html.twig', [
-            'controller_name' => 'EntrepriseController',
-            'entreprises' => $entreprises
+            'entreprises' => $entrepriseRepository->findAll(),
         ]);
     }
 
-    #[Route('/add', name: 'add')]
-    public function addEntreprise()
+    #[Route('/new', name: 'app_entreprise_new', methods: ['GET', 'POST'])]
+    public function new(Request $request, EntrepriseRepository $entrepriseRepository): Response
     {
-
-        return $this->render('entreprise/add.html.twig');
-    }
-
-    #[Route('/add/new', name: 'add_new', methods: 'POST')]
-    public function addNewEntreprise(EntrepriseRepository $entrepriseRepository)
-    {
-//        dd($_POST);
         $entreprise = new Entreprise();
-        $entreprise->setName($_POST['name']);
-        $entrepriseRepository->save($entreprise, true);
-        return $this->redirectToRoute('app_entreprise_index');
+        $form = $this->createForm(EntrepriseType::class, $entreprise);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $entrepriseRepository->save($entreprise, true);
+
+            return $this->redirectToRoute('app_entreprise_index', [], Response::HTTP_SEE_OTHER);
+        }
+
+        return $this->renderForm('entreprise/new.html.twig', [
+            'entreprise' => $entreprise,
+            'form' => $form,
+        ]);
     }
 
-    #[Route('/delete/{id}', name: 'delete')]
-    public function deleteEntreprise(Entreprise $entreprise, EntrepriseRepository $entrepriseRepository)
+    #[Route('/{id}', name: 'app_entreprise_show', methods: ['GET'])]
+    public function show(Entreprise $entreprise): Response
     {
-        $entrepriseRepository->remove($entreprise, true);
-        return $this->redirectToRoute('app_entreprise_index');
+        return $this->render('entreprise/show.html.twig', [
+            'entreprise' => $entreprise,
+        ]);
+    }
+
+    #[Route('/{id}/edit', name: 'app_entreprise_edit', methods: ['GET', 'POST'])]
+    public function edit(Request $request, Entreprise $entreprise, EntrepriseRepository $entrepriseRepository): Response
+    {
+        $form = $this->createForm(EntrepriseType::class, $entreprise);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $entrepriseRepository->save($entreprise, true);
+
+            return $this->redirectToRoute('app_entreprise_index', [], Response::HTTP_SEE_OTHER);
+        }
+
+        return $this->renderForm('entreprise/edit.html.twig', [
+            'entreprise' => $entreprise,
+            'form' => $form,
+        ]);
+    }
+
+    #[Route('/{id}', name: 'app_entreprise_delete', methods: ['POST'])]
+    public function delete(Request $request, Entreprise $entreprise, EntrepriseRepository $entrepriseRepository): Response
+    {
+        if ($this->isCsrfTokenValid('delete'.$entreprise->getId(), $request->request->get('_token'))) {
+            $entrepriseRepository->remove($entreprise, true);
+        }
+
+        return $this->redirectToRoute('app_entreprise_index', [], Response::HTTP_SEE_OTHER);
     }
 }
